Here we implemented a functionality for the players to be able to "stab" other players if they manage to land on the same posiiton the other player is.
Stabbing here basically means that if you catch up to a player who
was ahead of you and you land on the same box they landed, then you can
send the other player back to the starting posiiton at (0,0) effectively
giving them a huge disadvantage. The code first creates a  Constant variable
"STABBED". This is used to print the log which tells the user that
they have been stabbed by the other player and need to got back to the
origin. The code then checks using the worksflow object whose turn was it,
and based on that it changes the position of the  other player to (0,0).
Then a constant "RESTART_AGAIN" is  used to  declare to the other player that
they have been stabbed and need to restart the game again.

PS:- This commit also added a few comments to the original code to
help the programmer understand what is going on. Please pay no attention to
them.