Added a feature for the player to get a chance in case he gets a six, but limits how many 6 he can get.
Here we used a countSix variable to keep track of how many sixes the player
has rolled. Its value is limited to three at which point it will simply change
the players turn without actually calling "makeMove()".We also add a constant
variable GOT_SIX_THRICE which is used to print the log for the event.

This commit also adds the even "AFTER_MOVE" which tells the user his posiiton after
he makes his move. This functionality was implemented but not used.